+++ packages/agent_core/__init__.py
@@ -1,3 +1,19 @@
+from .models import AgentResponse, HRDomain, HRRequest, RetrievedSource, RoutingResult
 from .providers import MockProvider, ModelProvider, ProviderResponse
+from .retrieval import LocalKnowledgeRetriever
+from .router import ControllerRouter
+from .workflow import HRAgentWorkflow
 
-__all__ = ["MockProvider", "ModelProvider", "ProviderResponse"]
+__all__ = [
+    "AgentResponse",
+    "ControllerRouter",
+    "HRAgentWorkflow",
+    "HRDomain",
+    "HRRequest",
+    "LocalKnowledgeRetriever",
+    "MockProvider",
+    "ModelProvider",
+    "ProviderResponse",
+    "RetrievedSource",
+    "RoutingResult",
+]